[refactor] Move files to directory by its purpose
- Rename some files
