+++ .gitignore
@@ -1,0 +1,2 @@
+# 환경
+/ven
