+++ backend/routes/configs/db_models.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, Integer, String, ForeignKey
 from sqlalchemy.orm import relationship
 
-from database import Base
+from .db_connection import Base
 
 class User(Base):
     __tablename__ = "users"
