+++ frontend/pages/friends.html
